pyproject.toml
[project]
name = "flux-kontext-comfyui"
version = "1.0.0"
description = "ComfyUI node for Flux Kontext image editing"
authors = [
    {name = "Leonardo Miramondi", email = "leonardo.miramondi@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "Pillow>=8.0.0",
    "torch>=1.9.0",
    "numpy>=1.21.0"
]
keywords = ["comfyui", "flux", "kontext", "image-editing", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[project.urls]
Homepage = "https://github.com/leonardomiramondi/flux-kontext-comfyui"
Repository = "https://github.com/leonardomiramondi/flux-kontext-comfyui"
Issues = "https://github.com/leonardomiramondi/flux-kontext-comfyui/issues"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8"
]

[tool.comfyui]
title = "Flux Kontext Node"
description = "Image editing with Flux Kontext AI models"
category = "image/editing"
author = "Leonardo Miramondi"
tags = ["image", "editing", "ai", "flux", "kontext", "replicate"] 